added firebase config to server file
